+++ src/Service/CloudWatchLogs/src/Input/DescribeLogStreamsRequest.php
@@ -135,7 +135,7 @@ final class DescribeLogStreamsRequest extends Input
 
         // Prepare Body
         $bodyPayload = $this->requestBody();
-        $body = empty($bodyPayload) ? '{}' : json_encode($bodyPayload);
+        $body = empty($bodyPayload) ? '{}' : \json_encode($bodyPayload, 4194304);
 
         // Return the Request
         return new Request('POST', $uriString, $query, $headers, StreamFactory::create($body));