+++ src/Service/CognitoIdentityProvider/src/Input/RespondToAuthChallengeRequest.php
@@ -165,7 +165,7 @@ final class RespondToAuthChallengeRequest extends Input
 
         // Prepare Body
         $bodyPayload = $this->requestBody();
-        $body = empty($bodyPayload) ? '{}' : json_encode($bodyPayload);
+        $body = empty($bodyPayload) ? '{}' : \json_encode($bodyPayload, 4194304);
 
         // Return the Request
         return new Request('POST', $uriString, $query, $headers, StreamFactory::create($body));